+++ vector3.cpp
@@ -3,7 +3,7 @@
 
 vector3::vector3(){}
 
-vector3::vector3(float a,float b,float c)
+vector3::vector3(double a,double b,double c)
 {
   x=a;
   y=b;
@@ -20,7 +20,7 @@ vector3 vector3::operator-(vector3 r)
   return v(x-r.x, y-r.y, z-r.z);
 }
 
-vector3 vector3::operator*(float r)
+vector3 vector3::operator*(double r)
 {
   return v(x*r, y*r, z*r);
 }
@@ -36,7 +36,7 @@ vector3 vector3::operator&(vector3 r)
 }
 
 // Dot product
-float vector3::operator%(vector3 r)
+double vector3::operator%(vector3 r)
 {
   return x*r.x + y*r.y + z*r.z;
 }
@@ -53,7 +53,7 @@ vector3 vector3::operator!()
   return *this*(1/norm(*this));
 }
 
-float norm(vector3 a)
+double norm(vector3 a)
 {
   return sqrt(a%a);
 }
